09_APIs/FastAPI_template/requirements.txt
FastAPI
uvicorn
python-dotenv